services/service-suggest-friend-master/requirements.txt
opencv-python
numpy==1.21.5
pillow
requests
pymongo==4.3.3
python-dotenv
deepface
scipy
mtcnn
torch==2.0.1 
tensorflow-cpu==2.7.0 
protobuf==3.20.3
